Connect frontend to backend.
Now when a user signs in / up, the database will automatically be queried.
If the user is signing up, the db will be checked to see if their email is there or not. If not, user is created.
If the user is logging in, the db will be checked for their email and whether the associated password is correct.